Promenjeni longovi u integer

+++ uber-clone/src/main/java/com/backend/uberclone/model/User.java
@@ -8,9 +8,7 @@ import lombok.Setter;
 
 import javax.persistence.*;
 import java.sql.Timestamp;
-import java.util.HashSet;
 import java.util.List;
-import java.util.Set;
 
 @Setter
 @Getter
@@ -22,7 +20,7 @@ public abstract class User {
 
     @Id
     @GeneratedValue(strategy = GenerationType.TABLE)
-    private Long id;
+    private Integer id;
 
     @Column
     private String email;
